get all dynamic amount for dashboard checked

+++ src/main/java/com/example/crowdfunding/dto/ContributionSummaryDto.java
@@ -1,5 +1,6 @@
 package com.example.crowdfunding.dto;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 import com.example.crowdfunding.models.Utilisateur;
@@ -9,10 +10,10 @@ import lombok.Data;
 @Data
 public class ContributionSummaryDto {
     private Utilisateur utilisateur;
-    private int montantTotal;
+    private BigInteger montantTotal;
     private LocalDateTime derniereDate;
 
-    public ContributionSummaryDto(Utilisateur utilisateur, int montantTotal, LocalDateTime derniereDate) {
+    public ContributionSummaryDto(Utilisateur utilisateur, BigInteger montantTotal, LocalDateTime derniereDate) {
     this.utilisateur = utilisateur;
     this.montantTotal = montantTotal;
     this.derniereDate = derniereDate;
